@@ -2,15 +2,15 @@
 
 namespace EryseClient\Client\Administration\Profile\Voter;
 
-use EryseClient\Client\Profile\Entity\Profile;
+use EryseClient\Client\Profile\Entity\ProfileEntity;
 use EryseClient\Common\Voter\CrudVoter;
-use EryseClient\Server\User\Entity\User;
+use EryseClient\Server\User\Entity\UserEntity;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 
 /**
  * Class AdminProfileVoter
  *
- * @package EryseClient\Client\Administration\Profile\Voter
+ *
  */
 class AdminProfileVoter extends CrudVoter
 {
@@ -26,7 +26,7 @@ class AdminProfileVoter extends CrudVoter
             return false;
         }
 
-        if ($subject !== null && !$subject instanceof Profile) {
+        if ($subject !== null && !$subject instanceof ProfileEntity) {
             return false;
         }
 
@@ -44,7 +44,7 @@ class AdminProfileVoter extends CrudVoter
     {
         $user = $token->getUser();
 
-        if (!$user instanceof User) {
+        if (!$user instanceof UserEntity) {
             return false;
         }
 
@@ -59,12 +59,12 @@ class AdminProfileVoter extends CrudVoter
     }
 
     /**
-     * @param User $user
-     * @param Profile $userProfile
+     * @param UserEntity $user
+     * @param ProfileEntity $userProfile
      *
      * @return bool
      */
-    protected function canView(User $user, Profile $userProfile): bool
+    protected function canView(UserEntity $user, ProfileEntity $userProfile): bool
     {
         if ($this->userRoleService->isRoleAdmin($user)) {
             return true;
@@ -82,12 +82,12 @@ class AdminProfileVoter extends CrudVoter
     }
 
     /**
-     * @param User $user
-     * @param Profile $userProfile
+     * @param UserEntity $user
+     * @param ProfileEntity $userProfile
      *
      * @return bool
      */
-    protected function canEdit(User $user, Profile $userProfile): bool
+    protected function canEdit(UserEntity $user, ProfileEntity $userProfile): bool
     {
         if ($this->userRoleService->isRoleAdmin($user)) {
             return true;
