@@ -4,7 +4,7 @@ namespace EryseClient\Common\Service;
 
 /**
  * Class GdprService
- * @package EryseClient\Service
+ *
  */
 final class GdprService extends AbstractService
 {
