@@ -6,7 +6,7 @@ use Throwable;
 
 /**
  * Class InvalidTokenException
- * @package EryseClient\Exception\Profile
+ *
  */
 class ExpiredTokenException extends TokenException implements Throwable
 {
