@@ -6,7 +6,7 @@ use Throwable;
 
 /**
  * Class InvalidTokenException
- * @package EryseClient\Exception\Profile
+ *
  */
 class InvalidTokenException extends TokenException implements Throwable
 {
