@@ -6,24 +6,24 @@ use DateInterval;
 use DateTime;
 use EryseClient\Common\Service\AbstractService;
 use EryseClient\Common\Token\TokenInterface;
-use EryseClient\Server\Token\Entity\Token;
-use EryseClient\Server\Token\Entity\TokenType;
+use EryseClient\Server\Token\Entity\TokenEntity;
 use EryseClient\Server\Token\Exception\ExpiredTokenException;
 use EryseClient\Server\Token\Exception\InvalidTokenException;
 use EryseClient\Server\Token\Exception\TokenException;
 use EryseClient\Server\Token\Repository\TokenRepository;
-use EryseClient\Server\Token\Repository\TokenTypeRepository;
-use EryseClient\Server\User\Entity\User;
+use EryseClient\Server\Token\Type\Entity\TypeEntity;
+use EryseClient\Server\Token\Type\Repository\TypeRepository;
+use EryseClient\Server\User\Entity\UserEntity;
 use Exception;
 
 /**
  * Class TokenService
  *
- * @package EryseClient\Service
+ *
  */
 class TokenService extends AbstractService
 {
-    /** @var TokenTypeRepository $tokenTypeRepository */
+    /** @var TypeRepository $tokenTypeRepository */
     private $tokenTypeRepository;
 
     /** @var TokenRepository $tokenRepository */
@@ -32,22 +32,22 @@ class TokenService extends AbstractService
     /**
      * TokenService constructor.
      *
-     * @param TokenTypeRepository $tokenTypeRepository
+     * @param TypeRepository $tokenTypeRepository
      * @param TokenRepository $tokenRepository
      */
-    public function __construct(TokenTypeRepository $tokenTypeRepository, TokenRepository $tokenRepository)
+    public function __construct(TypeRepository $tokenTypeRepository, TokenRepository $tokenRepository)
     {
         $this->tokenTypeRepository = $tokenTypeRepository;
         $this->tokenRepository = $tokenRepository;
     }
 
     /**
-     * @param Token $token
+     * @param TokenEntity $token
      *
      * @return bool
      * @throws Exception
      */
-    public function isExpired(Token $token): bool
+    public function isExpired(TokenEntity $token): bool
     {
         $now = new DateTime();
 
@@ -55,19 +55,19 @@ class TokenService extends AbstractService
     }
 
     /**
-     * @param Token $token
+     * @param TokenEntity $token
      *
      * @return bool
      */
-    public function isValid(Token $token): bool
+    public function isValid(TokenEntity $token): bool
     {
         return !$token->isInvalid();
     }
 
     /**
-     * @return Token
+     * @return TokenEntity
      */
-    public function invalidate(Token $token): Token
+    public function invalidate(TokenEntity $token): TokenEntity
     {
         $token->setInvalid(true);
 
@@ -75,21 +75,21 @@ class TokenService extends AbstractService
     }
 
     /**
-     * @param User $user
+     * @param UserEntity $user
      *
-     * @return Token
+     * @return TokenEntity
      * @throws Exception
      */
-    public function getActivationToken(User $user): Token
+    public function getActivationToken(UserEntity $user): TokenEntity
     {
-        $token = new Token();
+        $token = new TokenEntity();
         $token->setHash($this->generateHash());
         $token->setUser($user);
         $createdAt = new DateTime();
         $expiresAt = new DateTime();
         $token->setCreatedAt($createdAt);
         $token->setExpiresAt($this->createExpirationDateTime($expiresAt, TokenInterface::EXPIRY_HOUR * 2));
-        $token->setType($this->tokenTypeRepository->findOneBy(['name' => TokenType::USER['ACTIVATION']]));
+        $token->setType($this->tokenTypeRepository->findOneBy(['name' => TypeEntity::USER['ACTIVATION']]));
 
         return $token;
     }
@@ -116,23 +116,23 @@ class TokenService extends AbstractService
     }
 
     /**
-     * @param User $user
+     * @param UserEntity $user
      *
-     * @return Token
+     * @return TokenEntity
      * @throws Exception
      */
-    public function generateResetPasswordToken(User $user): Token
+    public function generateResetPasswordToken(UserEntity $user): TokenEntity
     {
         $createdAt = new DateTime();
         $expiresAt = new DateTime();
 
-        $token = new Token();
+        $token = new TokenEntity();
         $token->setHash($this->generateHash());
         $token->setUser($user);
         $token->setCreatedAt($createdAt);
         $token->setExpiresAt($this->createExpirationDateTime($expiresAt, TokenInterface::EXPIRY_HOUR * 2));
         $token->setType(
-            $this->tokenTypeRepository->findOneBy(['name' => TokenType::USER['REQUEST_RESET_PASSWORD']])
+            $this->tokenTypeRepository->findOneBy(['name' => TypeEntity::USER['REQUEST_RESET_PASSWORD']])
         );
 
         return $token;
@@ -141,13 +141,13 @@ class TokenService extends AbstractService
     /**
      * Check whether the user has currently token of specified type and it is valid & un-expired
      *
-     * @param User $user
-     * @param TokenType $tokenType
+     * @param UserEntity $user
+     * @param TypeEntity $tokenType
      *
      * @return bool If Token of TokenType is found, is valid and is not expired
      * @throws Exception
      */
-    public function hasUserActiveTokenType(User $user, TokenType $tokenType): bool
+    public function hasUserActiveTokenType(UserEntity $user, TypeEntity $tokenType): bool
     {
         $result = count($this->tokenRepository->findUserValidTokensOfType($user, $tokenType));
 
@@ -155,14 +155,14 @@ class TokenService extends AbstractService
     }
 
     /**
-     * @param Token $token
+     * @param TokenEntity $token
      * @param string $tokenType
      *
      * @throws ExpiredTokenException
      * @throws InvalidTokenException
      * @throws TokenException
      */
-    public function verifyToken(Token $token, string $tokenType): void
+    public function verifyToken(TokenEntity $token, string $tokenType): void
     {
         if (!$token) {
             throw new TokenException(); //todo: add exception for missing token
